// frontend/src/api/types.ts
export interface components {
    schemas: {
        Booking: {
            id: string;
            eventTypeId: string;
            guestName: string;
            /** Format: email */
            guestEmail: string;
            /** Format: date-time */
            startTime: string;
            /** Format: date-time */
            endTime: string;
            /** Format: date-time */
            createdAt: string;
        };
        CreateBookingRequest: {
            eventTypeId: string;
            guestName: string;
            /** Format: email */
            guestEmail: string;
            /** Format: date-time */
            startTime: string;
        };
        CreateEventTypeRequest: {
            title: string;
            description: string;
            /** Format: int32 */
            durationMinutes: number;
        };
        ErrorResponse: {
            message: string;
            code: string;
        };
        EventType: {
            id: string;
            title: string;
            description: string;
            /** Format: int32 */
            durationMinutes: number;
        };
        Slot: {
            /** Format: date-time */
            startTime: string;
            /** Format: date-time */
            endTime: string;
        };
        UpdateBookingRequest: {
            guestName?: string;
            /** Format: email */
            guestEmail?: string;
            /** Format: date-time */
            startTime?: string;
        };
        UpdateEventTypeRequest: {
            title?: string;
            description?: string;
            /** Format: int32 */
            durationMinutes?: number;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}